RandomPhaseApproximation

Documentation for RandomPhaseApproximation.

Introduction

Standard random phase approximation (particle-hole channel) for quantum lattice systems based on the QuantumLattices and TightBindingApproximation packages.

Installation

In Julia v1.8+, please type ] in the REPL to use the package mode, then type this command:

pkg> add https://github.com/Quantum-Many-Body/RandomPhaseApproximation.jl

Getting Started

Examples of random phase approximation for quantum lattice system

Manuals

RandomPhaseApproximation.EigenRPA

RandomPhaseApproximation.EigenRPA

RandomPhaseApproximation.PHVertexMatrix

RandomPhaseApproximation.PHVertexMatrix

RandomPhaseApproximation.ParticleHoleSusceptibility

RandomPhaseApproximation.ParticleHoleSusceptibility

RandomPhaseApproximation.RPA

RandomPhaseApproximation.RPA

QuantumLattices.QuantumOperators.matrix

QuantumLattices.add!

RandomPhaseApproximation.chiq

RandomPhaseApproximation.chiq

RandomPhaseApproximation.chiq0

RandomPhaseApproximation.chiq0chiq

RandomPhaseApproximation.chiq0chiq

RandomPhaseApproximation.correlation

RandomPhaseApproximation.eigenrpa

RandomPhaseApproximation.fermifunc

RandomPhaseApproximation.isevenperm

RandomPhaseApproximation.issamesite

RecipesBase.apply_recipe

RecipesBase.apply_recipe

RecipesBase.apply_recipe

RandomPhaseApproximation.EigenRPA — Type

EigenRPA{P<:ReciprocalSpace, B<:BrillouinZone} <: Action

Eigen problem for standard random phase approximation.

RandomPhaseApproximation.EigenRPA — Method

EigenRPA(reciprocalspace::ReciprocalSpace, brillouinzone::BrillouinZone; eigvals_only::Bool=true, options...)

Construct a EigenRPA type. Attribute options contains (gauge=:icoordinate, exchange=false, η=1e-8, temperature=1e-12, μ=0.0, bands=nothing).

RandomPhaseApproximation.PHVertexMatrix — Type

PHVertexMatrix{D<:Number, Vq, Vk, T} <: MatrixRepresentation

Matrix representation of the particle-hole channel of two-body interaction terms:

\[\frac{1}{N}\sum_{k₁k₂q, \, \alpha\beta m n}[V^{ph}_{\alpha\beta, \, mn}(q)-V^{ph}_{\alpha m, \, \beta n}(k₂-k₁)]c^\dagger_{k₁-q, \, \alpha}c_{k₁, \, \beta}c^\dagger_{k₂, \, n}c_{k₂-q, \, m}\]

When the k₁ and k₂ are nothing, the exchange terms are omitted. Here, the Fourier transformation reads:

\[c^†_i = \frac{1}{\sqrt{N}} ∑_k c^†_k \exp(-i k rᵢ)\]

RandomPhaseApproximation.PHVertexMatrix — Method

PHVertexMatrix{D}(q, k₁, k₂, table, gauge::Symbol=:icoordinate) where {D<:Number}
PHVertexMatrix{D}(table, gauge::Symbol=:icoordinate)  where {D<:Number}
PHVertexMatrix{D}(q, table, gauge::Symbol=:icoordinate) where {D<:Number}

Get the matrix representation of particle-hole channel.

RandomPhaseApproximation.ParticleHoleSusceptibility — Type

ParticleHoleSusceptibility{P<:ReciprocalSpace, B<:BrillouinZone, E<:AbstractVector, S<:Operators} <: Action

Calculate the particle-hole susceptibility within random phase approximation.

Attribute options contains (η=0.01, gauge=:icoordinate, temperature=1e-12, μ=0.0, findk=false).

RandomPhaseApproximation.ParticleHoleSusceptibility — Method

ParticleHoleSusceptibility(reciprocalspace::ReciprocalSpace, brillouinzone::BrillouinZone, energies::Vector, operators::Tuple{Vector{<:Operators}, Vector{<:Operators}}; options...)

Construct a ParticleHoleSusceptibility type.

RandomPhaseApproximation.RPA — Type

RPA{L<:AbstractTBA, U<:RepresentationGenerator} <: Frontend

Random phase approximation in a fermionic system.

RandomPhaseApproximation.RPA — Method

RPA(tba::AbstractTBA, interactions::Tuple{Vararg{Term}}; neighbors::Union{Nothing, Int, Neighbors}=nothing)
RPA(lattice::AbstractLattice, hilbert::Hilbert, terms::Tuple{Vararg{Term}}, interactions::Tuple{Vararg{Term}}; neighbors::Union{Nothing, Int, Neighbors}=nothing)
RPA(tba::AbstractTBA{K, <:AnalyticalExpression}, hilbert::Hilbert, table::Table, interactions::Tuple{Vararg{Term}}; neighbors::Union{Nothing, Int, Neighbors}=nothing) where {K<:TBAKind}

Construct an RPA type.

QuantumLattices.QuantumOperators.matrix — Function

matrix(rpa::RPA, field::Symbol=:int; k=nothing, gauge=:icoordinate, kwargs...) -> Matrix

Get matrix of particle-hole channel of interaction.

QuantumLattices.add! — Method

add!(dest::AbstractMatrix, mr::PHVertexMatrix, m::Operator; kwargs...)

Get the matrix representation of an operator and add it to destination.

RandomPhaseApproximation.chiq — Method

chiq(vph::Array{<:Number, 3}, chi0::Array{ComplexF64, 4}) -> Array{ComplexF64, 4}

Get the susceptibility $χ_{αβ}(ω, q)$.

RandomPhaseApproximation.chiq — Method

chiq(eigvals::Vector{Vector{<:Number}}, eigvecs::Vector{Matrix{<:Number}}, o2bs::Vector{Matrix{ComplexF64}}, energies::AbstractVector; η::Float64=1e-2, imag_only::Bool=false) -> Array{ComplexF64, 4}

Get the $\chi_{ij, nm}(\omega, q)$. When imag_only is true, only the imaginary part is calculated.

Here, the eigenvalues, eigenvectors, and the orbital-to-band unitary matrices should be obtained by the method eigenrpa.

RandomPhaseApproximation.chiq0 — Method

chiq0(tba::AbstractTBA, brillouinzone::BrillouinZone, reciprocalspace::ReciprocalSpace, energies::Vector{Float64}; η::Float64=0.01, temperature::Float64=1e-12, μ::Float64=0.0, scflag=false, kwargs...) -> Array{Float64, 4}

Get the particle-hole susceptibilities χ⁰(ω, q).

Arguments

energies: the energy points

η: the magnitude of broaden

temperature: the temperature

μ: the chemical potential

scflag: false(default) for particle-hole channel, true for Nambu space

kwargs: the keyword arguments transferred to the matrix(tba; kwargs...) function

RandomPhaseApproximation.chiq0chiq — Method

chiq0chiq(
    eigvecs::Array{ComplexF64, 3}, eigvals::Array{Float64, 2}, brillouinzone::BrillouinZone, reciprocalspace::ReciprocalSpace, vph::Array{<:Number, 3}, energies::Vector{Float64};
    η::Float64=0.01, temperature::Float64=1e-12, μ::Float64=0.0, scflag=false
) -> Tuple{Array{ComplexF64, 4}, Array{ComplexF64, 4}}

Get the particle-hole susceptibilities χ⁰(ω, q) and χ(ω, q). The spectral function is satisfied by $A(ω, q) = \text{Im}[χ(ω+i0⁺, q)]$.

RandomPhaseApproximation.chiq0chiq — Method

chiq0chiq(
    tba::AbstractTBA, brillouinzone::AbstractVector{<:AbstractVector}, reciprocalspace::ReciprocalSpace, vph::Array{<:Number, 3}, energies::AbstractVector;
    η::Float64=0.01, temperature::Float64=1e-12, μ::Float64=0.0, scflag=false, kwargs...
) -> Tuple{Array{ComplexF64, 4}, Array{ComplexF64, 4}}

Get the particle-hole susceptibilities χ⁰(ω, q) and χ(ω, q). The spectral function is satisfied by $A(ω, q) = \text{Im}[χ(ω+i0⁺, q)]$.

Arguments

vph: the bare particle-hole vertex (vph[ndim, ndim, nq])

energies: the energy points

η: the magnitude of broaden

temperature: the temperature

μ: the chemical potential

scflag: false (default, no superconductivity), or true (BdG model)

kwargs: the keyword arguments transferred to the matrix(tba; kwargs...) function

RandomPhaseApproximation.correlation — Method

correlation(χ::Array{<:Number, 4}, reciprocalspace::ReciprocalSpace, operators::Tuple{Vector{<:Operators}, Vector{<:Operators}}, table; gauge=:rcoordinate) -> Matrix

Return physical particle-hole susceptibility.

RandomPhaseApproximation.eigenrpa — Method

eigenrpa(
    tba::AbstractTBA, brillouinzone::BrillouinZone, reciprocalspace::ReciprocalSpace, vph::Array{<:Number, 5};
    temperature::Float64=1e-12, μ::Float64=0.0, eigvals_only::Bool=true, η::Float64=1e-6, bands::Union{UnitRange{Int}, StepRange{Int,Int}, Vector{Int}, Nothing}=nothing, kwargs...
) -> Tuple{Array{Array{ComplexF64, 1}, 1}, Array{Matrix{ComplexF64}, 1}, Array{Matrix{ComplexF64}, 1}}

Get the eigenvalues, eigenvectors, and unitary transformation (from orbital to band) of particle-hole susceptibilities $χ_{αβ}(ω, k₁, k₂, q)$.

Now only the zero-temperature case is supported.

Arguments

vph: the particle-hole vertex, e.g. vph[ndim,ndim,nk,nk,nq] where sqrt(ndim) is the number of degrees of freedom in the unit cell, nk=length(brillouinzone), nq=length(reciprocalspace)

temperature: the temperature

μ: the chemical potential

eigvals_only: only the eigenvalues needed, when it is false the cholesky method is used

η: the small number to avoid the semi-positive Hamiltonian, i.e. Hamiltonian+diagm([η, η, ...])

bands: the selected bands to calculate the χ₀

kwargs: the keyword arguments transferred to the matrix(tba; kwargs...) function

RandomPhaseApproximation.fermifunc — Function

fermifunc(e::Real, temperature::Real=1e-12, μ::Real=0.0) -> Float64

Fermi distribution function. Boltzmann constant $k_B=1$.

RandomPhaseApproximation.isevenperm — Method

isevenperm(p::Vector) -> Bool

Judge the parity of permutations.

RandomPhaseApproximation.issamesite — Method

issamesite(op₁::CompositeIndex, op₂::CompositeIndex) -> Bool

Judge whether two composite indices are on site same site.

RecipesBase.apply_recipe — Function

@recipe plot(pack::Tuple{Algorithm{<:RPA}, Assignment{<:ParticleHoleSusceptibility}}, mode::Symbol=:χ)

Define the recipe for the visualization of particle-hole susceptibilities.

RecipesBase.apply_recipe — Function

@recipe plot(pack::Tuple{Algorithm{<:RPA}, Assignment{<:EigenRPA}})

Define the recipe for the visualization of particle-hole susceptibilities.

RecipesBase.apply_recipe — Function

@recipe plot(pack::Tuple{Algorithm{<:RPA}, Assignment{<:ParticleHoleSusceptibility}}, ecut::Float64, dE::Float64=1e-3, mode::Symbol=:χ, reim::Symbol=:re)